Remove all my code contributions
Due to the fact that it has taken so long for any activity to start on this repository, and i had asked numerous times when i was actively contributing for someone else to help out as well, I have decided to revoke my contributions.
Please do not add back any of my code as I will submit a DMCA takedown request.